@@ -21,17 +21,18 @@
     "Kirill Fomichev <fanatid@ya.ru> (https://github.com/fanatid)"
   ],
   "files": [
-    "lib",
-    "src",
+    "es",
+    "es5",
     "API.md",
     "binding.gyp",
+    "bindings.es",
     "bindings.js",
+    "elliptic.es",
     "elliptic.js",
-    "index.js",
+    "js.es",
     "js.js",
     "utils/has_lib.sh"
   ],
-  "main": "./index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/cryptocoinjs/secp256k1-node.git"
@@ -40,26 +41,31 @@
     "clean": "node-gyp clean",
     "coverage": "RANDOM_TESTS_REPEAT=1 nyc tape test/index.js",
     "coverage-lcov": "npm run coverage && nyc report -r lcov",
-    "install": "npm run rebuild || echo \"Secp256k1 bindings compilation fail. Pure JS implementation will be used.\"",
+    "install": "npm run rebuild",
     "lint": "standard",
     "rebuild": "node-gyp rebuild",
     "test": "npm run lint && npm run unit:node",
     "unit:browser": "karma start karma.conf.js",
-    "unit:node": "tape test/index.js |faucet"
+    "unit:node": "tape -r ./_require.js test/index.js | faucet"
   },
   "dependencies": {
     "bindings": "^1.2.1",
     "bip66": "^1.1.3",
     "bn.js": "^4.11.3",
     "create-hash": "^1.1.2",
-    "drbg.js": "^1.0.1",
     "elliptic": "^6.2.3",
     "nan": "^2.2.1"
   },
   "devDependencies": {
+    "babel-cli": "^6.16.0",
+    "babel-eslint": "^7.0.0",
+    "babel-plugin-transform-class-properties": "^6.18.0",
+    "babel-preset-es2015": "^6.16.0",
+    "babel-register": "^6.14.0",
     "bignum": "^0.11.0",
     "browserify": "^13.0.0",
     "faucet": "0.0.1",
+    "hash-test-vectors": "^1.3.2",
     "karma": "^0.13.22",
     "karma-browserify": "^5.0.4",
     "karma-chrome-launcher": "^0.2.3",
@@ -69,15 +75,19 @@
     "karma-tap": "^1.0.4",
     "node-gyp": "^3.3.1",
     "nyc": "^6.4.0",
-    "standard": "^7.0.0",
+    "require-extensions-order": "^1.0.1",
+    "standard": "*",
     "tape": "^4.5.1",
     "xorshift.js": "^1.0.3"
   },
   "engines": {
-    "node": ">=0.10"
+    "node": ">=4"
   },
-  "gypfile": true,
   "browser": {
     "./index.js": "./elliptic.js"
+  },
+  "gypfile": true,
+  "standard": {
+    "parser": "babel-eslint"
   }
 }
