@@ -8,161 +8,99 @@
 
 extern secp256k1_context* secp256k1ctx;
 
-NAN_METHOD(publicKeyCreate) {
+NAN_METHOD(public_key_create) {
   Nan::HandleScope scope;
 
-  v8::Local<v8::Object> private_key_buffer = info[0].As<v8::Object>();
-  CHECK_TYPE_BUFFER(private_key_buffer, EC_PRIVATE_KEY_TYPE_INVALID);
-  CHECK_BUFFER_LENGTH(private_key_buffer, 32, EC_PRIVATE_KEY_LENGTH_INVALID);
-  const unsigned char* private_key = (const unsigned char*) node::Buffer::Data(private_key_buffer);
-
-  unsigned int flags = SECP256K1_EC_COMPRESSED;
-  UPDATE_COMPRESSED_VALUE(flags, info[1], SECP256K1_EC_COMPRESSED, SECP256K1_EC_UNCOMPRESSED);
+  HANDLE_ARG_PRIVATE_KEY(0)
+  HANDLE_ARG_COMPRESSED(1)
 
   secp256k1_pubkey public_key;
   if (secp256k1_ec_pubkey_create(secp256k1ctx, &public_key, private_key) == 0) {
     return Nan::ThrowError(EC_PUBLIC_KEY_CREATE_FAIL);
   }
 
-  unsigned char output[65];
-  size_t output_length = 65;
-  secp256k1_ec_pubkey_serialize(secp256k1ctx, &output[0], &output_length, &public_key, flags);
-  info.GetReturnValue().Set(COPY_BUFFER(&output[0], output_length));
+  RETURN_PUBLIC_KEY(public_key)
 }
 
-NAN_METHOD(publicKeyConvert) {
+NAN_METHOD(public_key_convert) {
   Nan::HandleScope scope;
 
-  v8::Local<v8::Object> input_buffer = info[0].As<v8::Object>();
-  CHECK_TYPE_BUFFER(input_buffer, EC_PUBLIC_KEY_TYPE_INVALID);
-  CHECK_BUFFER_LENGTH2(input_buffer, 33, 65, EC_PUBLIC_KEY_LENGTH_INVALID);
-  const unsigned char* input = (unsigned char*) node::Buffer::Data(input_buffer);
-  size_t input_length = node::Buffer::Length(input_buffer);
+  HANDLE_ARG_PUBLIC_KEY(0)
+  HANDLE_ARG_COMPRESSED(1)
 
-  unsigned int flags = SECP256K1_EC_COMPRESSED;
-  UPDATE_COMPRESSED_VALUE(flags, info[1], SECP256K1_EC_COMPRESSED, SECP256K1_EC_UNCOMPRESSED);
-
-  secp256k1_pubkey public_key;
-  if (secp256k1_ec_pubkey_parse(secp256k1ctx, &public_key, input, input_length) == 0) {
-    return Nan::ThrowError(EC_PUBLIC_KEY_PARSE_FAIL);
-  }
-
-  unsigned char output[65];
-  size_t output_length = 65;
-  secp256k1_ec_pubkey_serialize(secp256k1ctx, &output[0], &output_length, &public_key, flags);
-  info.GetReturnValue().Set(COPY_BUFFER(&output[0], output_length));
+  RETURN_PUBLIC_KEY(public_key)
 }
 
-NAN_METHOD(publicKeyVerify) {
+NAN_METHOD(public_key_verify) {
   Nan::HandleScope scope;
 
-  v8::Local<v8::Object> input_buffer = info[0].As<v8::Object>();
-  CHECK_TYPE_BUFFER(input_buffer, EC_PUBLIC_KEY_TYPE_INVALID);
-  const unsigned char* input = (unsigned char*) node::Buffer::Data(input_buffer);
-  size_t input_length = node::Buffer::Length(input_buffer);
+  HANDLE_ARG_BUFFER(0, public_key_raw, EC_PUBLIC_KEY_TYPE_INVALID)
+  size_t public_key_raw_length = node::Buffer::Length(public_key_raw_buffer);
 
   secp256k1_pubkey public_key;
-  int result = secp256k1_ec_pubkey_parse(secp256k1ctx, &public_key, input, input_length);
-  info.GetReturnValue().Set(Nan::New<v8::Boolean>(result));
+  int result = secp256k1_ec_pubkey_parse(secp256k1ctx, &public_key, public_key_raw, public_key_raw_length);
+
+  RETURN_BOOLEAN(result)
 }
 
-NAN_METHOD(publicKeyTweakAdd) {
+NAN_METHOD(public_key_tweak_add) {
   Nan::HandleScope scope;
 
-  v8::Local<v8::Object> input_buffer = info[0].As<v8::Object>();
-  CHECK_TYPE_BUFFER(input_buffer, EC_PUBLIC_KEY_TYPE_INVALID);
-  CHECK_BUFFER_LENGTH2(input_buffer, 33, 65, EC_PUBLIC_KEY_LENGTH_INVALID);
-  const unsigned char* input = (unsigned char*) node::Buffer::Data(input_buffer);
-  size_t input_length = node::Buffer::Length(input_buffer);
-
-  v8::Local<v8::Object> tweak_buffer = info[1].As<v8::Object>();
-  CHECK_TYPE_BUFFER(tweak_buffer, TWEAK_TYPE_INVALID);
-  CHECK_BUFFER_LENGTH(tweak_buffer, 32, TWEAK_LENGTH_INVALID);
-  const unsigned char* tweak = (const unsigned char *) node::Buffer::Data(tweak_buffer);
-
-  unsigned int flags = SECP256K1_EC_COMPRESSED;
-  UPDATE_COMPRESSED_VALUE(flags, info[2], SECP256K1_EC_COMPRESSED, SECP256K1_EC_UNCOMPRESSED);
-
-  secp256k1_pubkey public_key;
-  if (secp256k1_ec_pubkey_parse(secp256k1ctx, &public_key, input, input_length) == 0) {
-    return Nan::ThrowError(EC_PUBLIC_KEY_PARSE_FAIL);
-  }
+  HANDLE_ARG_PUBLIC_KEY(0)
+  HANDLE_ARG_TWEAK(1)
+  HANDLE_ARG_COMPRESSED(2)
 
   if (secp256k1_ec_pubkey_tweak_add(secp256k1ctx, &public_key, tweak) == 0) {
     return Nan::ThrowError(EC_PUBLIC_KEY_TWEAK_ADD_FAIL);
   }
 
-  unsigned char output[65];
-  size_t output_length = 65;
-  secp256k1_ec_pubkey_serialize(secp256k1ctx, &output[0], &output_length, &public_key, flags);
-  info.GetReturnValue().Set(COPY_BUFFER(&output[0], output_length));
+  RETURN_PUBLIC_KEY(public_key)
 }
 
-NAN_METHOD(publicKeyTweakMul) {
+NAN_METHOD(public_key_tweak_mul) {
   Nan::HandleScope scope;
 
-  v8::Local<v8::Object> input_buffer = info[0].As<v8::Object>();
-  CHECK_TYPE_BUFFER(input_buffer, EC_PUBLIC_KEY_TYPE_INVALID);
-  CHECK_BUFFER_LENGTH2(input_buffer, 33, 65, EC_PUBLIC_KEY_LENGTH_INVALID);
-  const unsigned char* input = (unsigned char*) node::Buffer::Data(input_buffer);
-  size_t input_length = node::Buffer::Length(input_buffer);
-
-  v8::Local<v8::Object> tweak_buffer = info[1].As<v8::Object>();
-  CHECK_TYPE_BUFFER(tweak_buffer, TWEAK_TYPE_INVALID);
-  CHECK_BUFFER_LENGTH(tweak_buffer, 32, TWEAK_LENGTH_INVALID);
-  const unsigned char* tweak = (const unsigned char *) node::Buffer::Data(tweak_buffer);
-
-  unsigned int flags = SECP256K1_EC_COMPRESSED;
-  UPDATE_COMPRESSED_VALUE(flags, info[2], SECP256K1_EC_COMPRESSED, SECP256K1_EC_UNCOMPRESSED);
-
-  secp256k1_pubkey public_key;
-  if (secp256k1_ec_pubkey_parse(secp256k1ctx, &public_key, input, input_length) == 0) {
-    return Nan::ThrowError(EC_PUBLIC_KEY_PARSE_FAIL);
-  }
+  HANDLE_ARG_PUBLIC_KEY(0)
+  HANDLE_ARG_TWEAK(1)
+  HANDLE_ARG_COMPRESSED(2)
 
   if (secp256k1_ec_pubkey_tweak_mul(secp256k1ctx, &public_key, tweak) == 0) {
     return Nan::ThrowError(EC_PUBLIC_KEY_TWEAK_MUL_FAIL);
   }
 
-  unsigned char output[65];
-  size_t output_length = 65;
-  secp256k1_ec_pubkey_serialize(secp256k1ctx, &output[0], &output_length, &public_key, flags);
-  info.GetReturnValue().Set(COPY_BUFFER(&output[0], output_length));
+  RETURN_PUBLIC_KEY(public_key)
 }
 
-NAN_METHOD(publicKeyCombine) {
+NAN_METHOD(public_key_combine) {
   Nan::HandleScope scope;
 
-  v8::Local<v8::Array> input_buffers = info[0].As<v8::Array>();
-  CHECK_TYPE_ARRAY(input_buffers, EC_PUBLIC_KEYS_TYPE_INVALID);
-  CHECK_LENGTH_GT_ZERO(input_buffers, EC_PUBLIC_KEYS_LENGTH_INVALID);
+  v8::Local<v8::Array> public_keys_raw_buffers = info[0].As<v8::Array>();
+  CHECK_TYPE_ARRAY(public_keys_raw_buffers, EC_PUBLIC_KEYS_TYPE_INVALID);
+  CHECK_LENGTH_GT_ZERO(public_keys_raw_buffers, EC_PUBLIC_KEYS_LENGTH_INVALID);
+  unsigned int public_keys_count = public_keys_raw_buffers->Length();
 
-  unsigned int flags = SECP256K1_EC_COMPRESSED;
-  UPDATE_COMPRESSED_VALUE(flags, info[1], SECP256K1_EC_COMPRESSED, SECP256K1_EC_UNCOMPRESSED);
+  HANDLE_ARG_COMPRESSED(1)
 
-  std::unique_ptr<secp256k1_pubkey[]> public_keys(new secp256k1_pubkey[input_buffers->Length()]);
-  std::unique_ptr<secp256k1_pubkey*[]> ins(new secp256k1_pubkey*[input_buffers->Length()]);
-  for (unsigned int i = 0; i < input_buffers->Length(); ++i) {
-    v8::Local<v8::Object> public_key_buffer = v8::Local<v8::Object>::Cast(input_buffers->Get(i));
-    CHECK_TYPE_BUFFER(public_key_buffer, EC_PUBLIC_KEY_TYPE_INVALID);
-    CHECK_BUFFER_LENGTH2(public_key_buffer, 33, 65, EC_PUBLIC_KEY_LENGTH_INVALID);
+  std::unique_ptr<secp256k1_pubkey[]> public_keys(new secp256k1_pubkey[public_keys_count]);
+  std::unique_ptr<secp256k1_pubkey*[]> public_keys_ptrs(new secp256k1_pubkey*[public_keys_count]);
+  for (unsigned int i = 0; i < public_keys_count; ++i) {
+    v8::Local<v8::Object> public_key_raw_buffer = v8::Local<v8::Object>::Cast(public_keys_raw_buffers->Get(i));
+    CHECK_TYPE_BUFFER(public_key_raw_buffer, EC_PUBLIC_KEY_TYPE_INVALID);
+    CHECK_BUFFER_LENGTH2(public_key_raw_buffer, 33, 65, EC_PUBLIC_KEY_LENGTH_INVALID);
+    const unsigned char* public_key_raw = (unsigned char*) node::Buffer::Data(public_key_raw_buffer);
+    size_t public_key_raw_length = node::Buffer::Length(public_key_raw_buffer);
 
-    const unsigned char* input = (unsigned char*) node::Buffer::Data(public_key_buffer);
-    size_t input_length = node::Buffer::Length(public_key_buffer);
-    if (secp256k1_ec_pubkey_parse(secp256k1ctx, &public_keys[i], input, input_length) == 0) {
+    if (secp256k1_ec_pubkey_parse(secp256k1ctx, &public_keys[i], public_key_raw, public_key_raw_length) == 0) {
       return Nan::ThrowError(EC_PUBLIC_KEY_PARSE_FAIL);
     }
 
-    ins[i] = &public_keys[i];
+    public_keys_ptrs[i] = &public_keys[i];
   }
 
   secp256k1_pubkey public_key;
-  if (secp256k1_ec_pubkey_combine(secp256k1ctx, &public_key, ins.get(), input_buffers->Length()) == 0) {
+  if (secp256k1_ec_pubkey_combine(secp256k1ctx, &public_key, public_keys_ptrs.get(), public_keys_count) == 0) {
     return Nan::ThrowError(EC_PUBLIC_KEY_COMBINE_FAIL);
   }
 
-  unsigned char output[65];
-  size_t output_length = 65;
-  secp256k1_ec_pubkey_serialize(secp256k1ctx, &output[0], &output_length, &public_key, flags);
-  info.GetReturnValue().Set(COPY_BUFFER(&output[0], output_length));
+  RETURN_PUBLIC_KEY(public_key)
 }
