@@ -1,61 +1,59 @@
-'use strict'
-
 function genCombMulTo (alen, blen) {
   var len = alen + blen - 1
   var src = [
-    'var a = self.words;',
-    'var b = num.words;',
-    'var o = out.words;',
-    'var c = 0;',
-    'var lo;',
-    'var mid;',
-    'var hi;'
+    'const a = num1.words',
+    'const b = num2.words',
+    'const o = out.words',
+    'let c = 0',
+    'let lo',
+    'let mid',
+    'let hi'
   ]
-  for (var i = 0; i < alen; i++) {
-    src.push('var a' + i + ' = a[' + i + '] | 0;')
-    src.push('var al' + i + ' = a' + i + ' & 0x1fff;')
-    src.push('var ah' + i + ' = a' + i + ' >>> 13;')
+
+  for (let i = 0; i < alen; i++) {
+    src.push(`const a${i} = a[${i}] | 0`)
+    src.push(`const al${i} = a${i} & 0x1fff`)
+    src.push(`const ah${i} = a${i} >>> 13`)
   }
-  for (i = 0; i < blen; i++) {
-    src.push('var b' + i + ' = b[' + i + '] | 0;')
-    src.push('var bl' + i + ' = b' + i + ' & 0x1fff;')
-    src.push('var bh' + i + ' = b' + i + ' >>> 13;')
+  for (let i = 0; i < blen; i++) {
+    src.push(`const b${i} = b[${i}] | 0`)
+    src.push(`const bl${i} = b${i} & 0x1fff`)
+    src.push(`const bh${i} = b${i} >>> 13`)
   }
   src.push('')
-  src.push('out.length = ' + len + ';')
+  src.push(`out.length = ${len}`)
 
-  for (var k = 0; k < len; k++) {
-    var minJ = Math.max(0, k - alen + 1)
-    var maxJ = Math.min(k, blen - 1)
+  for (let k = 0; k < len; k++) {
+    const minJ = Math.max(0, k - alen + 1)
+    const maxJ = Math.min(k, blen - 1)
 
-    src.push('/* k = ' + k + ' */')
-    src.push('lo = Math.imul(al' + (k - minJ) + ', bl' + minJ + ');')
-    src.push('mid = Math.imul(al' + (k - minJ) + ', bh' + minJ + ');')
-    src.push('mid += Math.imul(ah' + (k - minJ) + ', bl' + minJ + ');')
-    src.push('hi = Math.imul(ah' + (k - minJ) + ', bh' + minJ + ');')
+    src.push(`/* k = ${k} */`)
+    src.push(`lo = Math.imul(al${k - minJ}, bl${minJ})`)
+    src.push(`mid = Math.imul(al${k - minJ}, bh${minJ})`)
+    src.push(`mid += Math.imul(ah${k - minJ}, bl${minJ})`)
+    src.push(`hi = Math.imul(ah${k - minJ}, bh${minJ})`)
 
-    for (var j = minJ + 1; j <= maxJ; j++) {
-      i = k - j
+    for (let j = minJ + 1; j <= maxJ; j++) {
+      let i = k - j
 
-      src.push('lo += Math.imul(al' + i + ', bl' + j + ');')
-      src.push('mid += Math.imul(al' + i + ', bh' + j + ');')
-      src.push('mid += Math.imul(ah' + i + ', bl' + j + ');')
-      src.push('hi += Math.imul(ah' + i + ', bh' + j + ');')
+      src.push(`lo += Math.imul(al${i}, bl${j})`)
+      src.push(`mid += Math.imul(al${i}, bh${j})`)
+      src.push(`mid += Math.imul(ah${i}, bl${j})`)
+      src.push(`hi += Math.imul(ah${i}, bh${j})`)
     }
 
-    src.push('var w' + k + ' = c + lo + ((mid & 0x1fff) << 13);')
-    src.push('c = hi + (mid >>> 13) + (w' + k + ' >>> 26);')
-    src.push('w' + k + ' &= 0x3ffffff;')
+    src.push(`let w${k} = c + lo + ((mid & 0x1fff) << 13)`)
+    src.push(`c = hi + (mid >>> 13) + (w${k} >>> 26)`)
+    src.push(`w${k} &= 0x3ffffff`)
   }
+
   // Store in separate step for better memory access
-  for (k = 0; k < len; k++) {
-    src.push('o[' + k + '] = w' + k + ';')
-  }
+  for (let k = 0; k < len; k++) src.push(`o[${k}] = w${k}`)
   src.push('if (c !== 0) {',
-           '  o[' + k + '] = c;',
-           '  out.length++;',
+           `  o[${len}] = c`,
+           '  out.length++',
            '}',
-           'return out;')
+           'return out')
 
   return src.join('\n')
 }
