@@ -1,19 +1,14 @@
-#!/usr/bin/env node
-'use strict'
+import { readFileSync, writeFileSync } from 'fs'
+import { join } from 'path'
 
-var fs = require('fs')
-var path = require('path')
+const sourceFilename = join(__dirname, '../src/messages.h')
+const outputFilename = join(__dirname, '../es/messages.json')
 
-var sourceFilename = path.join(__dirname, '../src/messages.h')
-var outputFilename = path.join(__dirname, '../lib/messages.json')
+const content = readFileSync(sourceFilename, 'utf-8')
+const messages = content.split('\n').reduce((result, line) => {
+  const match = line.match(/#define ([A-Z_0-9]+) "(.*)"/)
+  if (match !== null) result[match[1]] = match[2]
+  return result
+}, {})
 
-var result = {}
-var messages = fs.readFileSync(sourceFilename, 'utf-8').split('\n')
-for (var i = 0; i < messages.length; ++i) {
-  var match = messages[i].match(/#define ([A-Z_0-9]+) "(.*)"/)
-  if (match !== null) {
-    result[match[1]] = match[2]
-  }
-}
-
-fs.writeFileSync(outputFilename, JSON.stringify(result, null, 2) + '\n')
+writeFileSync(outputFilename, JSON.stringify(messages, null, 2) + '\n')
